+++ tests/tensorflow/experimental/data/reference_graphs/quantized/w_sym_ch_a_asym_t/resnet50.pb
