+++ tests/tensorflow/experimental/data/reference_graphs/quantized/w_sym_t_a_sym_t/simple.pb
