+++ tests/tensorflow/experimental/test_models/resnet.py
